relocated Header to App.js to make it central.
