+++ src/containers/countriesList.js
@@ -1,12 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { DISPLAYCOUNTRIES, RESET } from '../actions';
 import track from '../apis';
 import ListCountry from '../components/showCountries';
 import SearchBar from './searchBar';
-import Header from '../components/header';
-
 
 const mapDispatchToProps = dispatch => ({
   addCountries: countries => dispatch(DISPLAYCOUNTRIES(countries)),
@@ -21,14 +19,11 @@ const mapStateToProps = state => ({
 const CountriesList = ({
   addCountries, countries, filter, reset,
 }) => {
-  const [total, setTotal] = useState({});
-  const [style, setStyle] = useState({ display: 'block' });
   useEffect(() => {
     reset('');
     async function fetchData() {
       await track.countries()
         .then(resp => addCountries(resp));
-      await track.all().then(resp => setTotal(resp));
     }
     if (countries.length === 0) fetchData();
   }, [addCountries, reset, countries.length]);
@@ -39,17 +34,9 @@ const CountriesList = ({
     }
     return countries;
   };
-  const changeStyle = () => {
-    if (style.display === 'block') {
-      setStyle({ display: 'none' });
-    } else {
-      setStyle({ display: 'block' });
-    }
-  };
   const list = selected().map(country => <ListCountry country={country} key={country.country} />);
   return (
     <div>
-      <Header total={total} style={style} change={changeStyle} />
       <SearchBar />
       <div className="list">
         {list}
