+++ src/styles/marquee.css
@@ -6,9 +6,9 @@
     margin-bottom:10px;
     }
   .marquee  p {
-    -moz-animation: marquee 20s linear infinite;
-    -webkit-animation: marquee 20s linear infinite;
-    animation: marquee 20s linear infinite;
+    -moz-animation: marquee 15s linear infinite;
+    -webkit-animation: marquee 15s linear infinite;
+    animation: marquee 15s linear infinite;
     }
     @-moz-keyframes marquee {
     0% { transform: translateX(100%); }
